Update `compress-commons` for Node 15 support (#28)
* Update `compress-commons` for Node 15 support

Right now we're producing ZIP files with incorrect CRC32s on Node 15+.
This has a few of effects:

- macOS Finder gets confused and doesn't detect the ZIP file as password
  protected. Instead, it unzips everything as a 0 byte file

- Running `unzip` from the command line succeeds but warns about
  incorrect CRC32

- Our unit tests (correctly) fail

This is ultimately due to archiverjs/node-crc32-stream#31

To pull in the fix for that we need `crc32-stream@4.0.2`. That requires
updating both `compress-commons` and `zip-stream` to their latest
versions.

The version bumps look scary here but they're mostly due to dropping
support for old Node versions. They now require Node 10+; this should be
fine considering our Travis config only tests 12+.

* Update `package-lock.json`